requirements.txt
opencv_python==4.5.1.48
Pillow==8.2.0
